added Infinity to be a valid value for importance and location.unmetCost

+++ _prog/JsonVerifier/JsonVerifier/Models/Location.cs
@@ -3,5 +3,5 @@
 internal class Location
 {
     public int Id { get; set; }
-    public int UnmetCost { get; set; }
+    public double UnmetCost { get; set; }
 }
